pyproject.toml
[project]
name = "ytreza-dev"
version = "0.1.0"
description = ""
authors = [
    {name = "Jonathan",email = "johjo@ytreza.org"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest (>=8.4.1,<9.0.0)",
    "streamlit (>=1.47.0,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "types-requests (>=2.32.4.20250611,<3.0.0.0)"
]

[tool.poetry]
packages = [{include = "ytreza_dev", from = "src"}]


[tool.poetry.group.dev.dependencies]
approvaltests = "^15.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
python_files = [
    "test_*.py",
    "*_test.py",
    "tu_*.py",
    "ti_*.py",
    "ta_*.py",
]
